Fix build and test on Java 10+.
Fix the project so that it successfully builds and completes tests with Java 10+ and with current dependencies.
In the tests, move the source target forward to something still supported. Simplify the existing IGNORE list to handle the annotation processor failure within hickory with a single check that works for Java 9, 10, and 11 and is unlikely to change for future Java releases.
Update to a latest version of the parent pom to keep things current.
With the latest pom, it now runs findbugs. Fix the findbugs warning. For Java 9+ we will need to switch from findbugs to spotbugs to get it working fully, but that change is for a different PR. It should start in the parent pom.